requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
bson==0.5.10